requirements.txt
google-generativeai>=0.3.0
pandas>=1.5.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-mock>=3.10.0
